+++ tests/models/test_guild.py
@@ -19,7 +19,6 @@ class TestModelGuild:
             "motd": guild.motd,
             "show_links": guild.show_links,
             "voice_create": guild.voice_create,
-            "show_points": guild.show_points,
             "channels": [channel1.to_dict(), channel2.to_dict()],
             "awards": [award1.to_dict(), award2.to_dict()],
         }